+++ src/coremusic/audio/_base.py
@@ -75,7 +75,7 @@ class AudioFileLoaderMixin:
             import coremusic as cm
 
             with cm.AudioFile(str(self.audio_file)) as af:
-                self._audio_data = af.read_as_numpy()  # type: ignore[attr-defined]
+                self._audio_data = af.read_as_numpy()
                 self._sample_rate = af.format.sample_rate
 
         assert self._audio_data is not None